requirements.txt
APScheduler==3.11.2
blinker==1.9.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
numpy==2.4.1
pandas==2.3.3
pillow==12.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.9
six==1.17.0
tzdata==2025.3
tzlocal==5.3.1
Werkzeug==3.1.5
